Prevent docker disk usage from blowing up
We build six distinct variants of Alloy. The cross-product of this with
each benchmark program results in huge build directories of nearly
100GiB in size.

This paired with the directory-wide docker caching that buildx does is a
recipe for disaster because small changes can cause the docker image
directory to grow to multiples of this 100GiB. I learned this the hard
way by locking myself out of our benchmarking server due to a full disk!

This change uses multi-stage builds to throw away the build artefacts
after each significant step, which means that build images sizes never
exceed 30GiB and the final docker image with all programs built for
benchmarking is around 15GiB.