Flesh out new plotting code
After experimenting with different ways of presenting our results in the
paper, I've settled on an ideal presentation style using normalized
plots. Since results.py has become an unweidly mess, it's easier to just
copy over the bits I re-used and run the new plotting alongside it. This
will end up replacing it, but for now I don't want to break the build.